drone-sim-app/requirements.txt
streamlit
pygame
opencv-python-headless
numpy
Pillow